# mingw: improve performance of mingw_unlink()