@jeffhostetler 2020-04-20
See also https://github.com/git-for-windows/git/pull/2589