@derrickstolee 2020-04-20
@jeffhostetler just btw I ran this on the office repo with `git sparse-checkout set` alternating between every Word* directory (expand) and the empty set (contract):

|          |     | T1   | T2   | T3   | Avg      |          |
|----------|-----|------|------|------|----------|----------|
| expand   | old | 47.7 | 40.4 | 38.9 | 42.33333 |          |
| contract | old | 25.3 | 25.4 | 24.3 | 25       |          |
| expand   | new | 38.6 | 47.5 | 41.4 | 42.5     | -0.00394 |
| contract | new | 22.4 | 23.4 | 21.3 | 22.36667 | 0.105333 |

So this reduced the end-to-end time by 10% when deleting many files.