@jeffhostetler 2020-04-20
very nice.  thanks.  i'm going run a set after i finish getting my desktop put back together.

i'd also like to run a set of numbers on branch switching between representative L1's or L2's
or whatever the typical developer in Office does.